Prototype Favorite system and rework of selection panel code

+++ src/main/java/net/tomatentum/musicbot/command/commands/SearchCommand.java
@@ -1,12 +1,10 @@
 package net.tomatentum.musicbot.command.commands;
 
-import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.tomatentum.musicbot.MusicBot;
-import net.tomatentum.musicbot.command.utils.GuildCommand;
-import net.tomatentum.musicbot.music.SelectionPanel;
+import net.tomatentum.musicbot.utils.GuildCommand;
 
 import java.util.concurrent.TimeUnit;
 
@@ -14,17 +12,20 @@ public class SearchCommand implements GuildCommand {
 	@Override
 	public void execute(Member member, TextChannel channel, Message message, String[] args) {
 		message.delete().queue();
-		if (args.length > 1) {
-			StringBuilder URL = new StringBuilder();
+		if (channel.getGuild().getAudioManager().isConnected() && channel.getGuild().getAudioManager().getConnectedChannel().getMembers().contains(member) || !channel.getGuild().getAudioManager().isConnected()) {
 
-			for (int i = 1; i < args.length; i++) {
-				URL.append(args[i]).append(" ");
+			if (args.length > 1) {
+				StringBuilder URL = new StringBuilder();
+
+				for (int i = 1; i < args.length; i++) {
+					URL.append(args[i]).append(" ");
+				}
+				if (!channel.getGuild().getAudioManager().isConnected()) {
+					MusicBot.getInstance().getAudioManager().getMusicManager(channel.getGuild()).connect(member.getVoiceState().getChannel());
+				}
+				channel.sendMessage("🔍 Searching for: ``" + URL.toString() + "``").complete().delete().queueAfter(5, TimeUnit.SECONDS);
+				MusicBot.getInstance().getAudioManager().getMusicManager(channel.getGuild()).search(URL.toString(), channel);
 			}
-			if (!channel.getGuild().getAudioManager().isConnected()) {
-				MusicBot.getInstance().getAudioManager().getMusicManager(channel.getGuild()).connect(member.getVoiceState().getChannel());
-			}
-			channel.sendMessage("🔍 Searching for: ``" + URL.toString() + "``").complete().delete().queueAfter(5, TimeUnit.SECONDS);
-			MusicBot.getInstance().getAudioManager().getMusicManager(channel.getGuild()).search(URL.toString(), channel);
 		}
 	}
 }
